requirements.txt
apipkg==1.5
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
colormath==3.0.0
coverage==4.5.1
coveralls==1.3.0
cycler==0.10.0
Cython==0.29
decorator==4.3.0
docopt==0.6.2
execnet==1.5.0
flake8==3.5.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mkl-fft==1.0.6
mkl-random==1.0.1
mock==2.0.0
more-itertools==4.3.0
networkx==2.2
numpy==1.15.3
pandas==0.23.4
pbr==4.2.0
pluggy==0.8.0
py==1.7.0
pycodestyle==2.2.0
pyflakes==1.6.0
pyparsing==2.2.2
pytest==3.7.1
pytest-cov==2.5.1
pytest-forked==0.2
pytest-mock==1.10.0
pytest-xdist==1.22.5
python-dateutil==2.7.3
pytz==2018.6
PyVCF==0.6.8
PyYAML==3.13
requests==2.13.0
rpy2==2.9.4
scikit-learn==0.20.0
scipy==1.1.0
six==1.10.0
sklearn==0.0
tornado==5.1.1
urllib3==1.12
